+++ credmark/cmf/types/ledger_query.py
@@ -69,7 +69,16 @@ class LedgerQueryBase(contextlib.AbstractContextManager):
                 model_slug,
                 f'{offset=} needs to be a positive integer.')
 
-        aggregates_value = (None if aggregates is None
+        # Fix for contract ledger
+        # Customized columns needs to be converted to string to avoid losing precision.
+        # https://github.com/credmark/credmark-model-runner-api/issues/50
+        # pylint:disable=no-member
+        cols_customized = [(f'to_char({c})', c)
+                           for c in columns if c in self.bigint_cols]  # type: ignore
+        columns = [c for c in columns if c not in self.bigint_cols]  # type: ignore
+
+        aggregates = ([] if aggregates is None else aggregates) + cols_customized
+        aggregates_value = (None if len(aggregates) == 0
                             else [LedgerAggregate(expression=agg[0], asName=agg[1])
                                   for agg in aggregates])
 
@@ -98,7 +107,8 @@ class LedgerQuery(LedgerQueryBase):
                limit: Union[int, None] = None,
                offset: Union[int, None] = None,
                aggregates: Union[List[Tuple[str, str]], None] = None,
-               having: Union[str, None] = None) -> LedgerModelOutput:
+               having: Union[str, None] = None,
+               bigint_cols: Union[List[str], None] = None) -> LedgerModelOutput:
         """
         Query data from the table.
         """
@@ -113,6 +123,11 @@ class LedgerQuery(LedgerQueryBase):
                                             having=having)
 
         context = credmark.cmf.model.ModelContext.current_context()
-        return context.run_model(slug=self._cwgo_query,
-                                 input=model_input,
-                                 return_type=LedgerModelOutput)
+        ledger_out = context.run_model(slug=self._cwgo_query,
+                                       input=model_input,
+                                       return_type=LedgerModelOutput)
+        # pylint:disable=no-member, protected-access
+        ledger_out.set_bigint_cols(
+            self.bigint_cols +  # type: ignore
+            ([] if bigint_cols is None else bigint_cols))
+        return ledger_out